use macro to change fsck_error/fsck_fixes variable
define macro and use it to change variables for found error and fixed
count.

fsck_err_found() / fsck_err_failed() / fsck_err_canceled() / fsck_err_fixed()
fsck_err_found is called when fsck found error. fsck_err_fixed() is called when
error is fixed. fsck_err_canceled() is called when error turns out that is not
an error or does not need to be fixed.

Signed-off-by: JaeHoon Sim <jay.sim@lge.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>