ntfsck: check invalid attribute types in ntfs_mft_record_check()
When ntfsck check bytes_in_use field of mft entry by using attribute
length, There is infinite loop issue in there.
If attribute type and length is zero, space is not decreased and can not
escapse in loop. So this patch add the check of valid attribute type and
length.

Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>