ntfsck: don't recover if mft magic is magic_BAAD
If mft magic is invalid, It will be recovered in
ntfs_mft_record_check(). If it is magic_BAAD, It means fixup array may
be corrupted. then, such mft entry should be recovered with $MFTmirr or
removed. we can add the code that recover fixup array corruption in
ntfs_mst_post_read_fixup_warn() later.

Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>