ntfsck: optimize memory usage for lcn bitmap check
When fsck use th array to validate lcn bitmap(), Due to the clusters
of $MFTMirr that are very far behind located in patition, It is possible
to allocate unused memory in large partition that has the empty or small
number of files. This patch use the linked list and merge if lcn and length
is overapped with item in list.

Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>