ntfsck: optimize memory usage for lcn bitmap check
When fsck allocate memory to validate lcn bitmap($Bitmap), Due to
the clusters that are very far behind located in patition, It is possible
to allocate unused memory in large partition that has the empty or small
number of files. This patch use array indexing to optimize memory usage
in ntfsck.

Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>