ntfsck: fix checking index and adding orphan inode to lost+found
When orphaned inode is detected, ntfsck attempts to add it to its
parent.
Initially, ntfsck searches for the indexes of the parent inode
to determine if the index for orphaned inode exists or not.

If the index is found, ntfsck attempts to add orhpaned inode to
lost_found instead it's parent.
If the index is not found, ntfsck will try to add it to parent.
(If failed to add to parent, ntfsck will also try to add it to
lost+found.)

After adding it to any of them(parent or lost+found),
ntfsck will find index again and check inode.
(If failed to check inode, will be deleted.)

At this point, $FN fields for parent directory in orphaned inode and
index can be modified.
Mark dirty function call to make appropriate inode dirty.

Signed-off-by: JaeHoon Sim <jay.sim@lge.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>