ntfsck: fix infinite loop by circular parent number
When handling orphaned files, The parent of orphaned file can be
orphaned also. ntfsck add parent number to the orphan list to recover
the parent entry first. If parent number of child record is circluar,
infinite loop issue can happen. This patch lookup parent in orphan list
before adding it to the list. If there is same parent in list, simply
add it to lostfound.

Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>