ntfsck: fix '-C' option is not working when ntfs_mount is failed
'-C' is option to check volume dirty. It check flags after ntfs_mount().
So when ntfs_mount() is failed, It can not return proper value.

Modify it like below.
If '-C' option is given, set default fsck flags to check only.
And if ntfs_mount() is failed, ntfsck consider volume is dirty.

Signed-off-by: JaeHoon Sim <jay.sim@lge.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>