ntfsck: add condition to determine if inode is system file or not
When device is formatted on Windows, several system files are created.
Those files are sub-directory or files of $Extend.
But some files are not marked FILE_ATTR_SYSTEM, so it regarded as normal
file. unnecessary checking is occurred of system files.

To resolve it, add some condition to check system file.

Signed-off-by: JaeHoon Sim <jay.sim@lge.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>