ntfsck: fix too large peak memory issue
When store list entry for directory inode, directory inode is opening.
If there are so many sub-directories, A lot of directories should be
opening. It will cause too large peak memory issue. We change ntfs_inode
with mft_no in dir structure. When using it, directory inode should be
open.

Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>